requirements.txt
pandas>=2.0.0
plotly>=5.0.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0
numpy>=1.21.0
requests>=2.25.0
python-dotenv>=0.19.0